Ignore E2BIG caused by spawning with too many args.
This can happen on vanilla kernels with a small ARG_MAX constant. We
could alternatively create a temporary file for the list, but creating a
unique temp file is complex without library support, so the easy way is
to just ignore the error until our fix for objfsutil is released.

Change-Id: Ib1349c71a954c867bdc8f3e82e92b6cadb227c97