force tsserver to not scan google3 file systems
This is a quick-fix change to hopefully unwedge users.
Unfortunately it is difficult to test this, because the underlying cause of the
problem is related to using the tsserver under the checked in path, so modifying
the tsserver locally does not reproduce the problem.  So I think it's worth
trying a change like this to see if it helps.

This change was originally authored by evanm@ in cr/228738880.

Change-Id: I3ce86ba8bb96e71617171886748579e337233bc7