Fix a crash when transforming functions in modules.
When transforming a module declaration and block, a containing function
declarations parent pointer ends up pointing and the synthetic module
block and module declaration nodes.

If a function is contained in such a module declaration, and any of its
preceding statements are transformed, the `getOrCreateEmitNode`
triggered by `visitParameter` fails because `getSourceFileOfNode` fails
to find a SourceFile for the synthetic module declaration nodes.

This change avoids the crash by finding original nodes in
`getSourceFileOfNode`.